refactor demo for better custom usage
