+++ titiler/endpoints/factory_mosaic.py
@@ -395,13 +395,11 @@ class MosaicTilerFactory(BaseFactory):
             "/WMTSCapabilities.xml",
             response_class=XMLResponse,
             name=f"{self.router_prefix}wmts",
-            tags=["OGC"],
         )
         @self.router.get(
             "/{TileMatrixSetId}/WMTSCapabilities.xml",
             response_class=XMLResponse,
             name=f"{self.router_prefix}wmts",
-            tags=["OGC"],
         )
         def wmts(
             request: Request,
